+++ HISTORY.md
@@ -1,6 +1,18 @@
 # History
 ----
 
+## 9.15.0 / 2019-06-06
+
+- Add numeric keyboard support
+
+## 9.14.0 / 2019-05-24
+
+- Support Latvian locales
+
+## 9.13.0 / 2019-04-02
+
+- Support close onBlur
+
 ## 9.12.0 / 2019-03-16
 
 - RangerPicker support null as start or end